add sentence transformer to uploader

+++ scripts/build_docker.py
@@ -25,9 +25,6 @@ prompt = os.path.basename(os.environ['prompt'])
 version = open('version.md').read()
 
 # copy weights over to current folder for docker build
-if not os.path.exists(model):
-    print('[*] Copying weights to project root...')
-    shutil.copy(os.path.join(args.model_path, model), model)
 if not os.path.exists(tokenizer):
     print('[*] Copying tokenizer to project root...')
     shutil.copy(os.path.join(args.model_path, tokenizer), tokenizer)
@@ -36,6 +33,9 @@ if not os.path.isdir(sent_trans_model):
     shutil.copytree(os.path.join(args.model_path, sent_trans_model), sent_trans_model)
 
 if 'chat' in args.build:
+    if not os.path.exists(model):
+        print('[*] Copying weights to project root...')
+        shutil.copy(os.path.join(args.model_path, model), model)
     build_cmd = f'docker build -f Dockerfile --no-cache ' \
         f'--build-arg MODEL={model} --build-arg TOKENIZER={tokenizer} ' \
         f'--build-arg PROMPT={prompt} --build-arg SENT_TRANS_MODEL={sent_trans_model} ' \
@@ -46,14 +46,15 @@ if 'chat' in args.build:
 
 if 'upload' in args.build:
     build_cmd = f'docker build -f Dockerfile_upload --no-cache ' \
-        f'--build-arg TOKENIZER={tokenizer} ' \
+        f'--build-arg TOKENIZER={tokenizer} --build-arg SENT_TRANS_MODEL={sent_trans_model} ' \
         f'--build-arg SUDUAI_ENV={args.env} ' \
         f'-t asai-sudu:upload-{version} .'
     print(f'[*] Building upload app Docker image:\n{build_cmd}\n')
     subprocess.run(build_cmd, shell=True, text=True)
 
 # clean up
-os.remove(model)
+if 'chat' in args.build:
+    os.remove(model)
 os.remove(tokenizer)
-os.rmdir(sent_trans_model)
+shutil.rmtree(sent_trans_model)
